refactor : move scheduling helper test files from arbax to xlml
